requirements.txt
flask>=3.0
pyyaml>=6.0
Pillow>=10.0
httpx>=0.27
click>=8.0
